create errors without location (#4)
In our services we never create errors in place. We always create them
in error.go file in the package. Then when a function responds with
error it uses an error instance from error.go.

There is no reason to put location in that error, as it adds only
a clutter. We should be able to resolve only first Mask call.